V4

+++ vagrant/catalog/catalog.py
@@ -11,7 +11,7 @@ app = Flask(__name__)
 from sqlalchemy import create_engine
 from sqlalchemy import desc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, Item
+from database_setup import Base, Category, Item, User
 
 import random, string
 
@@ -24,11 +24,12 @@ import requests
 
 CLIENT_ID = json.loads(
     open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "My Catalog App"
 
 db = 'catalog'
 psql_user = 'vagrant'
 psql_pass = 'vagrant'
-conn_str = 'postgresql+psycopg2://{u}:{p}@localhost/catalog'.format(
+conn_str = 'postgresql+psycopg2://{u}:{p}@localhost/catalogwithusers'.format(
     u=psql_user, p=psql_pass)
 
 engine = create_engine(conn_str)
@@ -38,16 +39,15 @@ DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
 @app.route('/')
-@app.route('/index')
 def showMain():
-    return render_template('main.html')
+    return render_template('main.html', user_logged_in=isUserLogedIn(login_session))
 
-@app.route('/login')
+@app.route('/login/')
 def showLogin():
     state = ''.join(random.choice(
         string.ascii_uppercase + string.digits) for x in xrange(32))
     login_session['state'] = state
-    return render_template('login.html', STATE=state)
+    return render_template('login.html', STATE=state, user_logged_in=isUserLogedIn(login_session))
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
@@ -59,53 +59,53 @@ def gconnect():
     code = request.data
     try:
         # Upgrade the authorization code into a credentials object
-        oauth_flow = flow_from_clientsecrets('client_secrets.json',
-            scope='')
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
         oauth_flow.redirect_uri = 'postmessage'
-        credentials = oauth_flow.ste2_exchange(code)
+        credentials = oauth_flow.step2_exchange(code)
     except FlowExchangeError:
-        response = make_response(json.dumps(
-            'Failed to upgrade the authorization code.'), 401)
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
 
     # Check that the access token is valid
     access_token = credentials.access_token
-    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token={at}'.format(at=access_token))
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
     h = httplib2.Http()
     result = json.loads(h.request(url, 'GET')[1])
-    # If there was an error in the access token info, abort
+    # If there was an error in the access token info, abort.
     if result.get('error') is not None:
-        response = make_response(json.dumps(result.get('error')), 50)
+        response = make_response(json.dumps(result.get('error')), 500)
         response.headers['Content-Type'] = 'application/json'
-        return response
 
     # Verify that the access token is used for the intended user.
     gplus_id = credentials.id_token['sub']
     if result['user_id'] != gplus_id:
         response = make_response(
-            json.dumps("Token's user ID doesn't match giver user ID."), 401)
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
 
     # Verify that the access token is valid for this app.
     if result['issued_to'] != CLIENT_ID:
         response = make_response(
-            json.dumps("Token's client ID does not match app's one."), 401)
-        print "Token's client ID does not match app's one."
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
         response.headers['Content-Type'] = 'application/json'
         return response
 
     # Check to see if user is already logged in
-    stored_credentials = login_session.get('credentials')
+    stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
-    if stored_credentials is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'), 200)
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
     # Store the access token in the session for later use.
-    login_session['credentials'] = credentials
+    login_session['access_token'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
 
     # Get user info
@@ -113,11 +113,19 @@ def gconnect():
     params = {'access_token': credentials.access_token, 'alt': 'json'}
     answer = requests.get(userinfo_url, params=params)
 
+    data = json.loads(answer.text)
+
     login_session['username'] = data["name"]
     login_session['picture'] = data["picture"]
     login_session['email'] = data["email"]
 
-    poutput = ''
+    user_id = getUserId(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    
+    login_session['user_id'] = user_id
+
+    output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
     output += '!</h1>'
@@ -128,13 +136,43 @@ def gconnect():
     print "done!"
     return output
 
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session['access_token']
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: ' 
+    print login_session['username']
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+        del login_session['access_token'] 
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
 @app.route('/catalog/')
 def showCatalog():
     categories = session.query(Category).all()
     # only show last 10 items
     items = session.query(Item).order_by(desc('id')).limit(10)
-    return render_template('catalog.html', categories=categories,
-        items=items)
+    return render_template('catalog.html', categories=categories, items=items, user_logged_in=isUserLogedIn(login_session))
 
 @app.route('/catalog/c<int:category_id>/')
 def showItems(category_id):
@@ -142,39 +180,52 @@ def showItems(category_id):
     category = session.query(Category).filter_by(id = category_id).one()
     items = session.query(Item).filter_by(category_id = category_id).all()
     return render_template('category.html', categories=categories,
-        category=category, items=items)
+        category=category, items=items, user_logged_in=isUserLogedIn(login_session))
 
 @app.route('/catalog/i<int:item_id>/')
 def showItem(item_id):
     categories = session.query(Category).all()
     item = session.query(Item).filter_by(id = item_id).one()
-    return render_template('item.html', categories=categories, item=item)
+    creator = getUserInfo(item.user_id)
+    user_is_creator = False
+    if 'user_id' in login_session:
+        if creator.id == login_session['user_id']:
+            user_is_creator = True
+    return render_template('item.html', categories=categories, item=item, user_logged_in=isUserLogedIn(login_session), user_is_creator=user_is_creator)
 
 @app.route('/catalog/new/', methods=['GET', 'POST'])
 def newItem():
+    if 'username' not in login_session:
+        return redirect('/login')
+    categories = session.query(Category).all()
     if request.method == 'POST':
         result = session.query(Category.id).all()
         category_ids = [r[0] for r in result]
-        if int(request.form['cat']) not in category_ids:
-            flash("Creation of item failed! Not a valid category!")
-        else:
-            newItem = Item(name=request.form['name'],
-                description=request.form['desc'],
-                image=request.form['img'],
-                category_id=int(request.form['cat'])
-                )
-            session.add(newItem)
-            session.commit()
-            flash("New item created!")
+        newItem = Item(name=request.form['name'],
+            description=request.form['desc'],
+            image=request.form['img'],
+            category_id=request.form['cat'],
+            user_id=login_session['user_id']
+            )
+        session.add(newItem)
+        session.commit()
+        flash("New item created!")
         return redirect(url_for('showCatalog'))
     else:
-        return render_template('new_item.html')
+        return render_template('new_item.html', user_logged_in=isUserLogedIn(login_session),
+            categories=categories)
 
 @app.route('/catalog/i<int:item_id>/edit/', methods=['GET', 'POST'])
 def editItem(item_id):
-    '''Ren
-    '''
+    if 'username' not in login_session:
+        return redirect('/login')
+    categories = session.query(Category).all()
     item = session.query(Item).filter_by(id = item_id).one()
+    creator = getUserInfo(item.user_id)
+    if 'user_id' in login_session:
+        if creator.id != login_session['user_id']:
+            flash("You do not have permission to edit this item!")
+            return redirect(url_for('showItem', item_id=item.id))
     if request.method == 'POST':
         result = session.query(Category.id).all()
         category_ids = [r[0] for r in result]
@@ -194,22 +245,58 @@ def editItem(item_id):
             flash("Item edited successfully!")
         return redirect(url_for('showCatalog'))
     else:
-        return render_template('edit_item.html', item=item)
+        return render_template('edit_item.html', item=item, user_logged_in=isUserLogedIn(login_session),
+            categories=categories)
 
 @app.route('/catalog/i<int:item_id>/delete/', methods=['GET', 'POST'])
 def deleteItem(item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Item).filter_by(id = item_id).one()
+    creator = getUserInfo(item.user_id)
+    if 'user_id' in login_session:
+        if creator.id != login_session['user_id']:
+            flash("You do not have permission to delete this item!")
+            return redirect(url_for('showItem', item_id=item.id))
+            # OR THIS:
+            #return "<script>function myFunction() {alert('You are not authorized to delete this restaurant.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(item)
         session.commit()
         flash("Item {} deleted!".format(item.name))
         return redirect(url_for('showCatalog'))
     else:
-        return render_template('delete_item.html', item=item)
+        return render_template('delete_item.html', item=item, user_logged_in=isUserLogedIn(login_session))
 
+def isUserLogedIn(login_session):
+    if 'username' in login_session:
+        return True
+    else:
+        return False
+
+def getUserId(email):
+    try:
+        user = session.query(User).filter_by(email = email).one()
+        return user.id
+    except:
+        return None
+
+def getUserInfo(user_id):
+    print user_id
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+def createUser(login_session):
+    newUser = User(name = login_session['username'],
+        email = login_session['email'],
+        picture = login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email = login_session['email']).one()
+    return user.id
 
 if __name__ == '__main__':
     # set to False on production version
-    app.debug = True
+    app.debug = False
     app.secret_key = 'super_secret_key'
     app.run(host = '0.0.0.0', port = 8000)
